Added support for running the test suite using nose.
Couldn't find any instructions for running the tests, and ``python
setup.py test`` wasn't working for me.  Nose is a good one to use and
trivial to set up.

Added a requirements file for the testing dependencies (mock was missing)

Updated your README to have some instructions for potential
contributors.  You want it to be as easy as possible to go from zero to
all tests passing.